Update swap_alternate_element_of_linked_list_pairwise.cpp
First we will write a simple code of linked list which will have 2 functions:

1. insert() - which will insert elements in the linked list.

2. disp() - which will display the linked list.

Insert some values in the linked list.

Then make a function swap() which will swap alternate elements of linked list pairwise.

In the swap() function first we will make a pointer to node named temp which will initially points to the head node.

Then we will make a variable 'p' which we will use in swapping.

Now we will iterate the linked list such that the current node and the pointer to next node is not NULL which means 2 nodes exists and we will swap them. Otherwise if there exists only 1 node in the end then we will leave it as it is.

Next we will perform a swap operation between current and next node.

p=temp->data;

temp->data=temp->next->data;

temp->next->data=p;

1. In the above swapping code we have stored the data of current node in a p.

2. Then in the data of current node we have placed the data of next node. 'temp->next->data' will point to the data portion of next node.

3. Then in the data of next node we have placed the value of p.

In this way we have swapped the values of 2 alternate nodes pairwise.

At last we will double increment temp so that it will point to next pair of nodes (temp=temp->next->next).