Initial Commit
Program to find the median of two sorted arrays of same size and different size. Firstly, let us see what is median of the array?



Median is an element which divides the array into two parts - left and right. So the number of elements on the left side of the array will be equal or less than the number of elements on the right side. Now, let us consider the case of an array with odd number of elements.



Array = [9,11,16,7,2]

Sorted array = [2,7,9,11,16]



In this case, the median of this array is 9, since it divides the array into two parts: [2,7] and [11,16].



Further, let us consider the case of an array with even elements.



Array = [1,2,3,4,5,6]



In such a case, we will take the average between the last element of the left part and the first element of the right part. In this case, the median equals = (3 + 4) / 2 = 3.5